Customer_Segmentation/requirements.txt
Flask
pandas
numpy
scikit-learn
seaborn
matplotlib
kmodes
mysql-connector-python
